+++ src/commands/generators.ts
@@ -1,11 +1,17 @@
-import { Field, getFields, insertNewLinesParams } from "../utils/documentEdition";
+import { Field, getFields, insertNewLinesInHeaderParams, insertNewLinesInSourceParams } from "../utils/documentEdition";
 import { capitalize } from "../utils/string";
 import { AstNode, getAst } from "../lsp/ast";
 import { LSPContext } from "../lsp/setup";
-import { QuickPickOptions, window as Window } from "vscode";
-import { asConstRef, asConstRefIfNecessary, mustBeConstRef } from "../utils/type";
+import { QuickPickOptions, window as Window, workspace as Workspace } from "vscode";
+import { asConstRefIfNecessary } from "../utils/type";
+import { getSwitchSourceHeaderUri } from "./switchHeaderSource";
+import { LanguageClient } from "vscode-languageclient/node";
+import { openTextDocument } from "../utils/documents";
 
-type Content = string[];
+type Content = {
+    header: string[];
+    source: string[];
+};
 
 function getConstructorContent(className: string, fields: Field[]): Content {
     const paramsInside = fields.map((field) => `${asConstRefIfNecessary(field.type)} ${field.name}`).join(", ");
@@ -14,38 +20,89 @@ function getConstructorContent(className: string, fields: Field[]): Content {
         paramsOutside = " : ";
         paramsOutside += fields.map((field) => `${field.name}(${field.name})`).join(", ");
     }
-    return [`${className}(${paramsInside})${paramsOutside} {}`];
+    return {
+        header: [`${className}(${paramsInside});`],
+        source: ["", `${className}::${className}(${paramsInside})${paramsOutside} {}`, ""],
+    };
 }
 
 function getDestructorContent(className: string): Content {
-    return [`~${className}() {}`];
+    return {
+        header: [`~${className}();`],
+        source: ["", `${className}::~${className}() {}`, ""],
+    };
 }
 
 function getConstructorDestructorContent(className: string, fields: Field[]): Content {
-    return [...getConstructorContent(className, fields), ...getDestructorContent(className)];
+    const constructor = getConstructorContent(className, fields);
+    const destructor = getDestructorContent(className);
+
+    return {
+        header: [...constructor.header, ...destructor.header],
+        source: [...constructor.source, ...destructor.source],
+    };
 }
 
-function getGetters(fields: Field[]): Content {
-    return fields.map((field: Field) => {
+function getGetters(className: string, fields: Field[]): Content {
+    let source: string[] = [];
+    fields.forEach((field: Field) => {
         const { name, type } = field;
-        return `${asConstRefIfNecessary(type)} get${capitalize(name)}() const { return ${name}; }`;
+        source.push(
+            "",
+            `${asConstRefIfNecessary(type)} ${className}::get${capitalize(name)}() const`,
+            `{`,
+            `\treturn ${name};`,
+            `}`,
+            ""
+        );
     });
+
+    return {
+        header: fields.map((field: Field) => {
+            const { name, type } = field;
+            return `${asConstRefIfNecessary(type)} get${capitalize(name)}() const;`;
+        }),
+        source,
+    };
 }
 
-function getSetters(fields: Field[]): Content {
-    return fields.map((field: Field) => {
+function getSetters(className: string, fields: Field[]): Content {
+    let source: string[] = [];
+    fields.forEach((field: Field) => {
         const { name, type } = field;
-        return `void set${capitalize(name)}(${asConstRefIfNecessary(type)} ${name}) { this->${name} = ${name}; }`;
+        source.push(
+            "",
+            `void ${className}::set${capitalize(name)}(${asConstRefIfNecessary(type)} ${name})`,
+            `{`,
+            `\tthis->${name} = ${name};`,
+            `}`,
+            ""
+        );
     });
+
+    return {
+        header: fields.map((field: Field) => {
+            const { name, type } = field;
+            return `void set${capitalize(name)}(${asConstRefIfNecessary(type)} ${name});`;
+        }),
+        source,
+    };
 }
 
-function getGettersSetters(fields: Field[]): Content {
-    const getters = getGetters(fields);
-    const setters = getSetters(fields);
-    if (getters.length !== setters.length) return [];
-    const content = [];
-    for (let i = 0; i < getters.length; i++) content.push(getters[i], setters[i], "");
-    content.pop();
+function getGettersSetters(className: string, fields: Field[]): Content {
+    const getters = getGetters(className, fields);
+    const setters = getSetters(className, fields);
+
+    const content: Content = {
+        header: [],
+        source: [],
+    };
+
+    for (let i = 0; i < getters.header.length; i++) {
+        content.header.push("", getters.header[i], setters.header[i]);
+        content.source.push(...getters.source.slice(i * 6, i * 6 + 6), ...setters.source.slice(i * 6, i * 6 + 6));
+    }
+
     return content;
 }
 
@@ -62,11 +119,15 @@ async function showQuickPickFields(ast: AstNode): Promise<Field[]> {
     return fields.filter((field) => choices?.includes(field.name));
 }
 
-async function generateCode(context: LSPContext, getContent: (ast: AstNode) => Promise<Content>) {
+async function generateCode(
+    client: LanguageClient,
+    context: LSPContext,
+    getContent: (ast: AstNode) => Promise<Content>
+) {
     if (!Window.activeTextEditor) return;
 
     const editor = Window.activeTextEditor;
-    const document = editor.document;
+    const documentHeader = editor.document;
 
     const ast = await getAst(context);
     if (!ast) return;
@@ -74,41 +135,48 @@ async function generateCode(context: LSPContext, getContent: (ast: AstNode) => P
     const content = await getContent(ast);
     if (!content) return;
 
-    const [position, text] = insertNewLinesParams(ast, document, content, "public");
+    const [positionHeader, textHeader] = insertNewLinesInHeaderParams(ast, documentHeader, content.header, "public");
+    editor.edit((builder) => builder.insert(positionHeader, textHeader));
 
-    editor.edit((builder) => builder.insert(position, text));
+    const sourceUri = await getSwitchSourceHeaderUri(client);
+    const editorSource = await Window.showTextDocument(sourceUri);
+    const [positionSource, textSource] = insertNewLinesInSourceParams(ast, editorSource.document, content.source);
+    editorSource.edit((builder) => builder.insert(positionSource, textSource));
 }
 
 async function generateCodeWithFields(
+    client: LanguageClient,
     context: LSPContext,
     getContent: (ast: AstNode, fields: Field[]) => Promise<Content>
 ) {
-    generateCode(context, async (ast) => {
+    generateCode(client, context, async (ast) => {
         const fields = await showQuickPickFields(ast);
         return getContent(ast, fields);
     });
 }
 
-export async function generateConstructor(context: LSPContext) {
-    generateCodeWithFields(context, async (ast, fields) => getConstructorContent(ast.detail!, fields));
+export async function generateConstructor(client: LanguageClient, context: LSPContext) {
+    generateCodeWithFields(client, context, async (ast, fields) => getConstructorContent(ast.detail!, fields));
 }
 
-export async function generateDestructor(context: LSPContext) {
-    generateCode(context, async (ast) => getDestructorContent(ast.detail!));
+export async function generateDestructor(client: LanguageClient, context: LSPContext) {
+    generateCode(client, context, async (ast) => getDestructorContent(ast.detail!));
 }
 
-export async function generateConstructorDestructor(context: LSPContext) {
-    generateCodeWithFields(context, async (ast, fields) => getConstructorDestructorContent(ast.detail!, fields));
+export async function generateConstructorDestructor(client: LanguageClient, context: LSPContext) {
+    generateCodeWithFields(client, context, async (ast, fields) =>
+        getConstructorDestructorContent(ast.detail!, fields)
+    );
 }
 
-export async function generateGetters(context: LSPContext) {
-    generateCodeWithFields(context, async (_, fields) => getGetters(fields));
+export async function generateGetters(client: LanguageClient, context: LSPContext) {
+    generateCodeWithFields(client, context, async (ast, fields) => getGetters(ast.detail!, fields));
 }
 
-export async function generateSetters(context: LSPContext) {
-    generateCodeWithFields(context, async (_, fields) => getSetters(fields));
+export async function generateSetters(client: LanguageClient, context: LSPContext) {
+    generateCodeWithFields(client, context, async (ast, fields) => getSetters(ast.detail!, fields));
 }
 
-export async function generateGettersSetters(context: LSPContext) {
-    generateCodeWithFields(context, async (_, fields) => getGettersSetters(fields));
+export async function generateGettersSetters(client: LanguageClient, context: LSPContext) {
+    generateCodeWithFields(client, context, async (ast, fields) => getGettersSetters(ast.detail!, fields));
 }