+++ src/commands/switchHeaderSource.ts
@@ -1,4 +1,4 @@
-import { window as Window, workspace as Workspace, Uri } from "vscode";
+import { window as Window, workspace as Workspace, Uri, TextDocument } from "vscode";
 import { RequestType, TextDocumentIdentifier, LanguageClient } from "vscode-languageclient/node";
 import { openTextDocument } from "../utils/documents";
 
@@ -6,8 +6,8 @@ const switchHeaderSourceRequest = new RequestType<TextDocumentIdentifier, string
     "textDocument/switchSourceHeader"
 );
 
-export async function switchHeaderSource(client: LanguageClient): Promise<void> {
-    if (!Window.activeTextEditor) return;
+export async function getSwitchSourceHeaderUri(client: LanguageClient): Promise<Uri> {
+    if (!Window.activeTextEditor) return Promise.reject();
 
     const activePath = Window.activeTextEditor.document.fileName;
     const activeUri = Uri.file(activePath);
@@ -15,11 +15,15 @@ export async function switchHeaderSource(client: LanguageClient): Promise<void> 
     const destinationPath = await client.sendRequest(switchHeaderSourceRequest, activeDocumentIdentifier);
 
     if (!destinationPath) {
-        Window.showInformationMessage("Matching file not found.");
-        return;
+        Window.showInformationMessage("Matching header/source file not found.");
+        return Promise.reject();
     }
 
-    const destinationUri = Uri.parse(destinationPath);
+    return Uri.parse(destinationPath);
+}
+
+export async function switchHeaderSource(client: LanguageClient): Promise<void> {
+    const destinationUri = await getSwitchSourceHeaderUri(client);
     const document = await Workspace.openTextDocument(destinationUri);
 
     openTextDocument(document);