+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.8.3] - 2021-10-17
+
+- Change code actions order
+- Hide pickFields and getters/setters if no fields in class
+- Code actions inserts in source and header simultaneously
+
 ## [0.8.2] - 2021-10-11
 
 - Generate getters/setters with parameters